tMenuItemSlidingMenu ctor: drop the asm symbol rename
diffx/diffy were declared `int` with an __asm__("__20tMenuItemSlidingMenuUissssb")
label bridging the signature to retail's short-parameter symbol. Declaring them
`short` makes the natural cfront mangling produce that symbol directly, so the
rename is unnecessary -- measured identical, PASS 42/42 either way.

The `int` spelling came with the old device-laden body, whose comment claimed the
ABI needed SI-mode locals; that is no longer true of the current source and the
claim is dropped with it.

Both declaration sites updated (nfs4_types.h and the femenuoptions surface in
screendisplay_types.h); tMenuItemSlidingActivated already passed shorts.

  verify_asm: __20tMenuItemSlidingMenuUissssb PASS (42 insns)
  shared header touched, so both lanes rebuilt (gotcha #0 / #0b):
    build.py --skip-asm            0 SKIP
    build.py --out expected --no-link  0 SKIP
  objdiff: femenuoptions 92/92, screendisplay 3/3, tree 3289/3493
  femenudefs __15tGlobalMenuDefs is 196 diffs before and after -- pre-existing

+++ recon/frontend/common/femenuoptions.cpp
@@ -753,7 +753,7 @@ void tInsideBoxMenu::Draw(short x,short y,short w,short slideOffset,short)
 
 
 /* ---- tMenuItemSlidingMenu::ctor  [FEMENUOPTIONS.CPP:612-624] SLD-VERIFIED ---- */
-tMenuItemSlidingMenu::tMenuItemSlidingMenu(u_int textDescription,short width,short height,int diffx, int diffy,bool fillback)
+tMenuItemSlidingMenu::tMenuItemSlidingMenu(u_int textDescription,short width,short height,short diffx, short diffy,bool fillback)
   : tMenuItem(textDescription)
 {
   _vf = (void *)tMenuItemSlidingMenu_vtable;
